+++ apps/desktop/src/app/chat/index.tsx
@@ -759,8 +759,8 @@ const ChatViewContent = memo(function ChatViewContent({
               onAddUrl={onAddUrl}
               onAttachDroppedItems={onAttachDroppedItems}
               onAttachImageBlob={onAttachImageBlob}
-              onAttachPrCommentUrl={onAttachPrCommentUrl}
               onAttachPastedText={onAttachPastedText}
+              onAttachPrCommentUrl={onAttachPrCommentUrl}
               onCancel={onCancel}
               onPasteClipboardImage={onPasteClipboardImage}
               onPickFiles={onPickFiles}
